refactor: extract admin SelectionProvider to shared location
